refactor: remove legacy command definitions
